camera qualify fixed, median software more testing, restored num detectors to 2

+++ DE1_SOC_D8M_RTL/software/slave_de1/cameraTest.h
@@ -2,9 +2,12 @@
 #define CAMERATEST_H
 
 #include <stdio.h>
+#include <stdint.h>
 #include "camera.h"
 
 int cameraTest() {
+    // uint1
+    // printf("getCoordsFromHW(0), ")
     srand(time(NULL));
     struct CameraInterface *camera_interface = CameraInterface_new();
     for (int i = 0; i < BUFFER_SIZE; i++) {
@@ -13,15 +16,15 @@ int cameraTest() {
     printf("Buffer:\n");
     for (int i = 0; i < NUM_DETECTORS; i++) {
         for (int j = 0; j < BUFFER_SIZE; j++) {
-            printf("%f %f\n", camera_interface->buf[i][j][0], camera_interface->buf[i][j][1]);
+            printf("%i %i\n", camera_interface->buf[i][j][0], camera_interface->buf[i][j][1]);
         }
     }
-    float median[NUM_DETECTORS * DIMENSIONS];
+    int median[NUM_DETECTORS * DIMENSIONS];
     getMedian(camera_interface, median);
     printf("Median:\n");
     for (int i = 0; i < NUM_DETECTORS; i++) {
         for (int j = 0; j < DIMENSIONS; j++) {
-            printf("%f ", median[i * DIMENSIONS + j]);
+            printf("%i ", median[i * DIMENSIONS + j]);
         }
         printf("\n");
     }
